Use MinHeap to improve entity id recycling
- Create MinHeap datastructure

Previously, the the id of the last entity that was despawned was used for the next entity. But this was not necessarily the lowest unused entity id available and could therefore lead to unnecessarily large component pools and allocatitons.
Improved now by always using the lowest entity id available by using a MinHeap datastructure